backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
backcall==0.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
Django==3.1.7
django-cors-headers==3.7.0
django-extensions==3.1.1
django-filter==2.4.0
django-multiselectfield==0.1.12
django-rest-passwordreset==1.1.0
djangorestframework==3.12.2
drf-yasg==1.20.0
filelock==3.0.12
flex==6.14.1
idna==2.10
importlib-metadata==3.10.0
inflection==0.5.1
ipython==7.20.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonpointer==2.0
kiwisolver==1.3.1
llvmlite==0.36.0
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
matrix-factorization==1.3
numba==0.53.1
numpy==1.20.2
packaging==20.9
pandas==1.2.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.16
pycparser==2.20
Pygments==2.8.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rfc3987==1.3.8
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
strict-rfc3339==0.7
threadpoolctl==2.1.0
traitlets==5.0.5
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
validate-email==1.3
virtualenv==20.4.2
wcwidth==0.2.5
zipp==3.4.1
